Deprecate Yoast_Feature_Toggles
